+++ packages/moonstone/Button/Button.js
@@ -202,12 +202,6 @@ const ButtonBaseFactory = factory({css: componentCss}, ({css}) =>
 		},
 
 		render: ({children, disabled, icon, ...rest}) => {
-			// Do not add the ARIA attribute if the selected prop is omitted to avoid the potentially
-			// confusing readout for the common case of a standalone Button or IconButton.
-			if ('selected' in rest) {
-				rest['aria-pressed'] = rest.selected;
-			}
-
 			delete rest.backgroundOpacity;
 			delete rest.color;
 			delete rest.minWidth;