+++ packages/moonstone/CHANGELOG.md
@@ -6,13 +6,20 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Deprecated
 
+- `moonstone/Scroller` props `horizontal` and `vertical`. Deprecated props are replaced with `direction` prop. `horizontal` and `vertical` will be removed in 2.0.0.
+
 ### Added
+
 - `moonstone/Image` support for `children` prop inside images
+- `moonstone/Scroller` prop `direction` which replaces `horizontal` and `vertical` props.
 
 ### Changed
 - `moonstone/VideoPlayer` to set play/pause icon to display "play" when rewinding or fast forwarding
+- `moonstone/VideoPlayer` to rewind or fast forward when previous command is slow-forward or slow-rewind respectively
+- `moonstone/VideoPlayer` to fast forward when previous command is slow-forward and it reaches the last of its play rate
 
 ### Fixed
+
 - `moonstone/GridListImageItem` to center its selection icon on the image instead of the item
 
 - `moonstone/Input` to have correct `Tooltip` position in `RTL`